Remove `Impl` behaviour and explicitly use ETS implementation
